LPD-78127 Split data-set types into specific files
