+++ projects/js-toolkit/packages/js-api/data-set/cell-renderer.ts
@@ -1,0 +1,19 @@
+/**
+ * SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ */
+
+/**
+ * Public type contracts for Frontend Data Set (FDS) custom cell
+ * renderers: the HTML element builder a renderer implements to draw a
+ * table cell.
+ */
+
+export interface FDSTableCellHTMLElementBuilderArgs {
+	itemData?: Record<string, unknown>;
+	value: boolean | number | string | object | [];
+}
+
+export interface FDSTableCellHTMLElementBuilder {
+	(args: FDSTableCellHTMLElementBuilderArgs): HTMLElement;
+}
