+++ projects/js-toolkit/packages/js-api/data-set/filter.ts
@@ -1,0 +1,39 @@
+/**
+ * SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ */
+
+/**
+ * Public type contracts for Frontend Data Set (FDS) custom filters: HTML
+ * element builders for rendering, OData query builders for server-side
+ * filtering, and description builders for human-readable filter
+ * summaries, all composed by `FDSFilter`.
+ */
+
+export interface FDSFilterData<T> {
+	selectedData: T;
+}
+
+export interface FDSFilterHTMLElementBuilderArgs<T> {
+	fieldName?: string;
+	filter: FDSFilterData<T>;
+	setFilter: (partialFilter: Partial<FDSFilterData<T>>) => void;
+}
+
+export interface FDSFilterHTMLElementBuilder<T> {
+	(args: FDSFilterHTMLElementBuilderArgs<T>): HTMLElement;
+}
+
+export interface FDSFilterODataQueryBuilder<T> {
+	(selectedData: T): string;
+}
+
+export interface FDSFilterDescriptionBuilder<T> {
+	(selectedData: T): string;
+}
+
+export interface FDSFilter<T> {
+	descriptionBuilder: FDSFilterDescriptionBuilder<T>;
+	htmlElementBuilder: FDSFilterHTMLElementBuilder<T>;
+	oDataQueryBuilder: FDSFilterODataQueryBuilder<T>;
+}
